allow 'xri://$xrds*simple' Type to be filtered out
Some XRDS consumers (notably spring.io/spring-security) are choking on
the XRDS-Simple Type element that was getting included in every XRD
block.  This change allows that type to be filtered out using the
'xrds_simple' WordPress filter.

This change also means that the XRDS-Simple Type is only being included
in the 'main' XRD block, and that any others will manually need to
include it themselves if they care about it.

Fixes #3